Update sfdb_update_cohorts.py
A little bug: there seems to be strange behavior in the clone method. 

When there are no updates to the files/folders in the working project and the archiving script is run multiple times, the summaryfile.txt files only appear in the first archived folder. i.e. if I ran the archiving script twice without editing anything in the contributor folders, two archived subfolders will be generated, but only the first will be populated with the relevant files and folders, while the second will have empty folders. 

When I tested by adding a new random file into the (same, unedited) working project and ran the archiving script, an archived subfolder was created which only contained the new random file but no summaryfile.txt files.