Squashed commit of the following: Script for aggregating summaryfile.txt files for multiple cohorts.
Two files are being committed: a YML configuration file and a python script for updating cohorts.

The YML file contains the project IDs of the Working Project, Archive Project, and all Contributor Projects.

The python script uses the project IDs in the YML file to:
(1) archive all files and folders in the Working Project into the Archive Project, and
(2) update the Working Project with cohort data from all Contributor Projects.

The folder structure of the Archive, Working, and Contributor Projects, and other information can be found in the wiki.